added side bar w/ login, signin shows welcome img
