+++ eventally-app/src/App.css
@@ -1,38 +1,75 @@
 .App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
+  width: 100vw;
+  height: 100vh;
   display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
+  flex-direction: row;
+  /* justify-content: space-between; */
+}
+
+body {
+  margin: 0;
+  padding: 0;
+}
+
+.Sidebar {
+  height: 100%;
+  width: 250px;
+  background-color: #0b427a;
+}
+
+.SidebarList {
+  height: auto;
+  padding: 80px 0;
+  width: 100%;
+}
+
+img {
+  width: 200px;
+  padding: 21px;
+}
+
+.Main {
+  width: 55%;
+}
+
+.Main img {
+  width: 95%;
+}
+
+.SidebarList .row {
+  width: 100%;
+  height: 60px;
+  list-style-type: none;
+  margin: 0%;
+  display: flex;
+  flex-direction: row;
   color: white;
+  justify-content: center;
+  align-items: center;
+  font-family: "DM Sans", sans-serif;
 }
 
-.App-link {
-  color: #61dafb;
+.SidebarList .row:hover {
+  cursor: pointer;
+  background-color: #2e9eb7;
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+.row #icon {
+  flex: 30%;
+  display: grid;
+  place-items: center;
+}
+
+.row #title {
+  flex: 70%;
+}
+
+.Calendar2 {
+  place-items: end;
+  max-width: 80% !important;
+}
+
+.mbsc-eventcalendar {
+  width: 30%;
+  text-align: right;
 }
